keypaior bug fix

+++ eks/create-eks.sh
@@ -30,12 +30,18 @@ curl --silent --location "https://github.com/weaveworks/eksctl/releases/download
 sudo mv /tmp/eksctl /usr/local/bin
 
 echo Create a keypair
-cd ~
-if [ ! -f ~/${keypairName}.pem ]; then
-  aws ec2 create-key-pair --key-name $keypairName --region $region --query 'KeyMaterial' --output text > ${keypairName}.pem
-  chmod 400 ~/${keypairName}.pem
-  sleep 10
+
+aws ec2 describe-key-pairs --key-names $keypairName | grep $keypairName
+if [[ $? -eq 1 ]]; then
+  echo "Keypair does not exists, try creating one"
+  cd ~
+  if [ ! -f ~/${keypairName}.pem ]; then
+    aws ec2 create-key-pair --key-name $keypairName --region $region --query 'KeyMaterial' --output text > ${keypairName}.pem
+    chmod 400 ~/${keypairName}.pem
+    sleep 10
+  fi
 fi
+
 
 cd ~
 eksctl create cluster --node-private-networking --ssh-access --ssh-public-key $keypairName --name eks-fabric --region $region --node-type m5.xlarge --node-volume-size 200 --kubeconfig=./kubeconfig.eks-fabric.yaml
@@ -77,17 +83,19 @@ echo Running ~/aws-eks-fabric/efs/deploy-ec2.sh - this will use CloudFormation t
 cd ~/aws-eks-fabric/
 ./efs/deploy-ec2.sh
 
-PublicDnsNameBastion=$(aws ec2 describe-instances --region $region --filters "Name=tag:Name,Values=EFS FileSystem Mounted Instance" "Name=instance-state-name,Values=running" | jq '.Reservations | .[] | .Instances | .[] | .PublicDnsName' | tr -d '"')
+PublicDnsNameBastion=$(aws ec2 describe-instances --region $region --filters "Name=tag:Name,Values=eks-fabric-bastion" "Name=instance-state-name,Values=running" | jq '.Reservations | .[] | .Instances | .[] | .PublicDnsName' | tr -d '"')
 echo public DNS of EC2 bastion host: $PublicDnsNameBastion
 
 PrivateDnsNameEKSWorker=$(aws ec2 describe-instances --region $region --filters "Name=tag:Name,Values=eks-fabric-*-Node" "Name=instance-state-name,Values=running" | jq '.Reservations | .[] | .Instances | .[] | .PrivateDnsName' | tr -d '"')
 echo private DNS of EKS worker nodes, accessible from Bastion only since they are in a private subnet: $PrivateDnsNameEKSWorker
-cd ~
-# we need the keypair on the bastion, since we can only access the K8s worker nodes from the bastion
-scp -i ${keypairName}.pem -q ~/${keypairName}.pem  ec2-user@${PublicDnsNameBastion}:/home/ec2-user/${keypairName}.pem
+if [ ! -f ~/${keypairName}.pem ]; then
+  cd ~
+  scp -i ${keypairName}.pem -q ~/${keypairName}.pem  ec2-user@${PublicDnsNameBastion}:/home/ec2-user/${keypairName}.pem
 
-echo Prepare the EC2 bastion for use by copying the kubeconfig and aws config and credentials files from Cloud9
-cd ~
-scp -i ${keypairName}.pem -q ~/kubeconfig.eks-fabric.yaml  ec2-user@${PublicDnsNameBastion}:/home/ec2-user/kubeconfig.eks-fabric.yaml
-scp -i ${keypairName}.pem -q ~/.aws/config  ec2-user@${PublicDnsNameBastion}:/home/ec2-user/config
-scp -i ${keypairName}.pem -q ~/.aws/credentials  ec2-user@${PublicDnsNameBastion}:/home/ec2-user/credentials
+  echo Prepare the EC2 bastion for use by copying the kubeconfig and aws config and credentials files from Cloud9
+  cd ~
+  scp -i ${keypairName}.pem -q ~/kubeconfig.eks-fabric.yaml  ec2-user@${PublicDnsNameBastion}:/home/ec2-user/kubeconfig.eks-fabric.yaml
+  scp -i ${keypairName}.pem -q ~/.aws/config  ec2-user@${PublicDnsNameBastion}:/home/ec2-user/config
+  scp -i ${keypairName}.pem -q ~/.aws/credentials  ec2-user@${PublicDnsNameBastion}:/home/ec2-user/credentials
+fi
+echo Success creating EKS cluster
